Headless browsing
From now on, the code won't open a browser and will do what it does in headless mode. This should save some cpu usage.

For debugging purposes in the future, you can always make changes to the function that returns a headless driver to return a non-headless one.

Or you could make the code save screenshots and check those screenshots.